小修改

+++ test/plac/audio.c
@@ -112,8 +112,8 @@ int main() {
   dctx->callback         = play_audio;
   dctx->userdata         = dctx;
 
-  u16 samplerate;
-  u32 nsamples;
+  u32 samplerate;
+  u64 nsamples;
   plac_read_header(dctx, &samplerate, &nsamples);
 
   snd_pcm_open(&pcm_out, "default", SND_PCM_STREAM_PLAYBACK, 0);
